backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.1.0
transformers>=4.36.0
bitsandbytes>=0.41.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
